omap3: Unify omap2_set_globals_3[43,6x]x functions
The only difference between them is the physical address of the
uart4 port, which is only present in 36xx chips.

We don't really need to care about keeping these 2 functions, since
the decision to use uart4 is more cleanly done later when we do have
access to omap_revision variable.

Signed-off-by: Sergio Aguirre <saaguirre@ti.com>
Acked-by: Kevin Hilman <khilman@deeprootsystems.com>
[tony@atomide.com: added comment for the uart4_phys]
Signed-off-by: Tony Lindgren <tony@atomide.com>

Conflicts:
	arch/arm/mach-omap2/board-3630sdp.c
	arch/arm/mach-omap2/board-devkit8000.c
	arch/arm/mach-omap2/board-omap3stalker.c
	arch/arm/mach-omap2/board-omap3touchbook.c
	arch/arm/mach-omap2/board-rx51.c
	arch/arm/mach-omap2/board-zoom3.c
	arch/arm/plat-omap/common.c
	arch/arm/plat-omap/include/plat/common.h